pyproject.toml
[tool.poetry]
name = "ddns-update"
version = "0.1.0"
description = ""
authors = ["Sam Gunner <me@sgunner.me>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
